# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agmux API",
    "version": "current",
    "description": "Current agmux HTTP API surface used by UI and agents."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4821"
    }
  ],
  "paths": {
...
    "/api/agent-sessions/{provider}/{sessionId}/conversation": {
      "get": {
        "summary": "Read conversation messages for an agent session",
        "parameters": [
          { "$ref": "#/components/parameters/provider" },
          { "$ref": "#/components/parameters/sessionId" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad request" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/agent-sessions/{provider}/{sessionId}/restore": {
      "post": {
        "summary": "Restore a provider session into a new PTY",
        "parameters": [
          { "$ref": "#/components/parameters/provider" },
          { "$ref": "#/components/parameters/sessionId" }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Restored" },
          "400": { "description": "Bad request" },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Server error" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "provider": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "sessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    }
  }
}